Refactor: Move CustomScrollbar to core component
